+++ src/javascript/js-1/compute.ts
@@ -1,0 +1,5 @@
+import pipe from './utils/pipe';
+import rectanglify from './utils/rectanglify';
+import maxItemAssociation from './utils/maxItemAssociation';
+
+export const compute = pipe(rectanglify, maxItemAssociation);