+++ src/javascript/js-1/utils/pipe/pipe.ts
@@ -1,0 +1,8 @@
+/**
+ * Очень простая реализация функции pipe
+ *
+ * @param {...any[]} fns
+ */
+const pipe = (...fns: any[]) => (x: any) => fns.reduce((v, f) => f(v), x);
+
+export default pipe;