+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.1.1 - 2021-01-31
+
+Функциональность версии основывается на следующей задаче:
+
+**1 месяц. Введение в курс Modern JavaScript Frameworks. Домашнее задание 1**
+
+Merge request:
+
+([MR-4](https://github.com/taksenov/aksenov-timofey-otus/pull/4))
+
+### Добавлено
+
+- Функциональность: утилита `pipe` для запуска функций слева направо
+  `src/javascript/js-1/utils/pipe`
+- Функциональность: утилита `rectanglify` для проверки входных данных и создания
+  прямоугольной матрицы `src/javascript/js-1/utils/rectanglify`
+- Функциональность: утилита `maxItemAssociation` для опредения максималотнго
+  списка рекомендаций `src/javascript/js-1/utils/maxItemAssociation`
+- Функциональность: набор тестов для проверки работоспособности
+  `src/javascript/js-1/compute.test.ts`, а так же набор мокапов данных для
+  тестирования
+
 ## 0.0.1 - 2021-01-30
 
 ### Добавлено